Add opportunity for implementation-specific initialization
It may be desirable for concrete implementations to add their own
initialization logic. The known use-case is creation of a lookup map
based on an element of the definition.

To avoid the "Call Super" anti-pattern, an empty protected `prepare`
method has been added. It is always called at the end of initialization.
This gives implementations the opportunity to override the method
to add their own initialization logic.

This change also includes miscellaneous semantic and documentation
improvements.